Update the step-by-step tutorials
