+++ tutorials/nservicebus-step-by-step/3-multiple-endpoints/Solution/Sales/PlaceOrderHandler.cs
@@ -3,22 +3,15 @@ using Messages;
 using Microsoft.Extensions.Logging;
 using NServiceBus;
 
-namespace Sales
+namespace Sales;
+
+public class PlaceOrderHandler(ILogger<PlaceOrderHandler> logger) :
+    IHandleMessages<PlaceOrder>
 {
-    public class PlaceOrderHandler :
-        IHandleMessages<PlaceOrder>
+
+    public Task Handle(PlaceOrder message, IMessageHandlerContext context)
     {
-        private readonly ILogger<PlaceOrderHandler> logger;
-
-        public PlaceOrderHandler(ILogger<PlaceOrderHandler> logger)
-        {
-            this.logger = logger;
-        }
-
-        public Task Handle(PlaceOrder message, IMessageHandlerContext context)
-        {
-            logger.LogInformation("Received PlaceOrder, OrderId = {orderId}", message.OrderId);
-            return Task.CompletedTask;
-        }
+        logger.LogInformation("Received PlaceOrder, OrderId = {orderId}", message.OrderId);
+        return Task.CompletedTask;
     }
 }
