+++ tutorials/nservicebus-step-by-step/3-multiple-endpoints/Solution/Sales/Program.cs
@@ -1,31 +1,26 @@
 ﻿using System;
 using System.Threading.Tasks;
 using NServiceBus;
+using Microsoft.Extensions.Hosting;
 
-namespace Sales
+namespace Sales;
+
+class Program
 {
-    #region SalesProgram
-
-    class Program
+    static async Task Main(string[] args)
     {
-        static async Task Main()
-        {
-            Console.Title = "Sales";
+        Console.Title = "Sales";
 
-            var endpointConfiguration = new EndpointConfiguration("Sales");
-            // Choose JSON to serialize and deserialize messages
-            endpointConfiguration.UseSerialization<SystemJsonSerializer>();
+        var builder = Host.CreateApplicationBuilder(args);
 
-            var transport = endpointConfiguration.UseTransport<LearningTransport>();
+        var endpointConfiguration = new EndpointConfiguration("Sales");
 
-            var endpointInstance = await Endpoint.Start(endpointConfiguration);
+        endpointConfiguration.UseSerialization<SystemJsonSerializer>();
 
-            Console.WriteLine("Press Enter to exit.");
-            Console.ReadLine();
+        endpointConfiguration.UseTransport<LearningTransport>();
+        
+        builder.UseNServiceBus(endpointConfiguration);
 
-            await endpointInstance.Stop();
-        }
+        await builder.Build().RunAsync();
     }
-
-    #endregion
 }
