+++ tutorials/nservicebus-step-by-step/4-publishing-events/Solution/Billing.Messages/OrderBilled.cs
@@ -1,10 +1,9 @@
 ﻿using NServiceBus;
 
-namespace Messages
+namespace Messages;
+
+public class OrderBilled :
+    IEvent
 {
-    public class OrderBilled :
-        IEvent
-    {
-        public string OrderId { get; set; }
-    }
+    public string OrderId { get; set; }
 }
