+++ tutorials/nservicebus-step-by-step/4-publishing-events/Solution/Sales.Messages/PlaceOrder.cs
@@ -1,10 +1,9 @@
 ﻿using NServiceBus;
 
-namespace Messages
+namespace Messages;
+
+public class PlaceOrder :
+    ICommand
 {
-    public class PlaceOrder :
-        ICommand
-    {
-        public string OrderId { get; set; }
-    }
+    public string OrderId { get; set; }
 }
