+++ tutorials/nservicebus-step-by-step/4-publishing-events/Solution/Sales/Program.cs
@@ -1,28 +1,26 @@
 ﻿using System;
 using System.Threading.Tasks;
+using Microsoft.Extensions.Hosting;
 using NServiceBus;
 
-namespace Sales
+namespace Sales;
+
+class Program
 {
-    class Program
+    static async Task Main(string[] args)
     {
-        static async Task Main()
-        {
-            Console.Title = "Sales";
+        Console.Title = "Sales";
 
-            var endpointConfiguration = new EndpointConfiguration("Sales");
+        var builder = Host.CreateApplicationBuilder(args);
 
-            // Choose JSON to serialize and deserialize messages
-            endpointConfiguration.UseSerialization<SystemJsonSerializer>();
+        var endpointConfiguration = new EndpointConfiguration("Sales");
 
-            var transport = endpointConfiguration.UseTransport<LearningTransport>();
+        endpointConfiguration.UseSerialization<SystemJsonSerializer>();
 
-            var endpointInstance = await Endpoint.Start(endpointConfiguration);
+        endpointConfiguration.UseTransport<LearningTransport>();
 
-            Console.WriteLine("Press Enter to exit.");
-            Console.ReadLine();
+        builder.UseNServiceBus(endpointConfiguration);
 
-            await endpointInstance.Stop();
-        }
+        await builder.Build().RunAsync();
     }
 }
