+++ tutorials/nservicebus-step-by-step/4-publishing-events/Solution/Shipping/OrderPlacedHandler.cs
@@ -5,16 +5,10 @@ using NServiceBus;
 
 namespace Shipping
 {
-    public class OrderPlacedHandler :
+    public class OrderPlacedHandler(ILogger<OrderPlacedHandler> logger) :
         IHandleMessages<OrderPlaced>
     {
-        private readonly ILogger<OrderPlacedHandler> logger;
-
-        public OrderPlacedHandler(ILogger<OrderPlacedHandler> logger)
-        {
-            this.logger = logger;
-        }
-
+        
         public Task Handle(OrderPlaced message, IMessageHandlerContext context)
         {
             logger.LogInformation("Received OrderPlaced, OrderId = {orderId} - Should we ship now?",message.OrderId);
