+++ tutorials/nservicebus-step-by-step/5-retrying-errors/Solution/Billing.Messages/OrderBilled.cs
@@ -1,10 +1,9 @@
 ﻿using NServiceBus;
 
-namespace Messages
+namespace Messages;
+
+public class OrderBilled :
+    IEvent
 {
-    public class OrderBilled :
-        IEvent
-    {
-        public string OrderId { get; set; }
-    }
+    public string OrderId { get; set; }
 }
