+++ tutorials/nservicebus-step-by-step/5-retrying-errors/Solution/Sales.Messages/OrderPlaced.cs
@@ -1,10 +1,9 @@
 ﻿using NServiceBus;
 
-namespace Messages
+namespace Messages;
+
+public class OrderPlaced :
+    IEvent
 {
-    public class OrderPlaced :
-        IEvent
-    {
-        public string OrderId { get; set; }
-    }
+    public string OrderId { get; set; }
 }
