+++ tutorials/nservicebus-step-by-step/5-retrying-errors/Solution/Sales/PlaceOrderHandler.cs
@@ -1,44 +1,35 @@
 ﻿using System.Threading.Tasks;
 using Messages;
 using NServiceBus;
+using System;
+using Microsoft.Extensions.Logging;
 
-namespace Sales
+namespace Sales;
+
+public class PlaceOrderHandler(ILogger<PlaceOrderHandler> logger) :
+    IHandleMessages<PlaceOrder>
 {
-    using System;
-    using Microsoft.Extensions.Logging;
+    static Random random = new Random();
 
-    public class PlaceOrderHandler :
-        IHandleMessages<PlaceOrder>
+    public Task Handle(PlaceOrder message, IMessageHandlerContext context)
     {
-        private readonly ILogger<PlaceOrderHandler> logger;
+        logger.LogInformation("Received PlaceOrder, OrderId = {orderId}", message.OrderId);
 
-        public PlaceOrderHandler(ILogger<PlaceOrderHandler> logger)
+        // This is normally where some business logic would occur
+
+        // Uncomment to test throwing a systemic exception
+        //throw new Exception("BOOM");
+
+        // Uncomment to test throwing a transient exception
+        //if (random.Next(0, 5) == 0)
+        //{
+        //    throw new Exception("Oops");
+        //}
+
+        var orderPlaced = new OrderPlaced
         {
-            this.logger = logger;
-        }
-
-        static Random random = new Random();
-
-        public Task Handle(PlaceOrder message, IMessageHandlerContext context)
-        {
-            logger.LogInformation("Received PlaceOrder, OrderId = {orderId}", message.OrderId);
-
-            // This is normally where some business logic would occur
-
-            // Uncomment to test throwing a systemic exception
-            //throw new Exception("BOOM");
-
-            // Uncomment to test throwing a transient exception
-            //if (random.Next(0, 5) == 0)
-            //{
-            //    throw new Exception("Oops");
-            //}
-
-            var orderPlaced = new OrderPlaced
-            {
-                OrderId = message.OrderId
-            };
-            return context.Publish(orderPlaced);
-        }
+            OrderId = message.OrderId
+        };
+        return context.Publish(orderPlaced);
     }
 }
