+++ tutorials/nservicebus-step-by-step/5-retrying-errors/Solution/Shipping/OrderPlacedHandler.cs
@@ -3,22 +3,15 @@ using Messages;
 using Microsoft.Extensions.Logging;
 using NServiceBus;
 
-namespace Shipping
+namespace Shipping;
+
+public class OrderPlacedHandler(ILogger<OrderPlacedHandler> logger) :
+    IHandleMessages<OrderPlaced>
 {
-    public class OrderPlacedHandler :
-        IHandleMessages<OrderPlaced>
+       
+    public Task Handle(OrderPlaced message, IMessageHandlerContext context)
     {
-        private readonly ILogger<OrderPlacedHandler> logger;
-
-        public OrderPlacedHandler(ILogger<OrderPlacedHandler> logger)
-        {
-            this.logger = logger;
-        }
-
-        public Task Handle(OrderPlaced message, IMessageHandlerContext context)
-        {
-            logger.LogInformation("Received OrderPlaced, OrderId = {orderId} - Should we ship now?", message.OrderId);
-            return Task.CompletedTask;
-        }
+        logger.LogInformation("Received OrderPlaced, OrderId = {orderId} - Should we ship now?", message.OrderId);
+        return Task.CompletedTask;
     }
 }
